is this a endianess issue?

+++ waterbox/ares64/ares/ares/n64/si/si.cpp
@@ -236,7 +236,7 @@ auto SI::scan() -> void {
             if(addressCRC(address) == (n5)input[2]) {
               for(u32 index : range(recv - 1)) {
                 output[index] = transferPak.read(address++);
-                printf("%04X, %02X\n", address - 1, output[index]);
+                printf("%04X, %02X\n", address - 1, (u32)output[index]);
               }
               output[recv - 1] = dataCRC({&output[0], recv - 1});
               valid = 1;
@@ -279,7 +279,7 @@ auto SI::scan() -> void {
             if(addressCRC(address) == (n5)input[2]) {
               for(u32 index : range(send - 3)) {
                 transferPak.write(address++, input[3 + index]);
-                printf("%04X, %02X\n", address - 1, input[3 + index]);
+                printf("%04X, %02X\n", address - 1, (u32)input[3 + index]);
               }
               output[0] = dataCRC({&input[3], send - 3});
               valid = 1;
